ref: Restructure commands file structure and importing process
