+++ src/commands/debug_files/check.rs
