+++ src/commands/files/delete.rs
