+++ src/commands/projects/mod.rs
@@ -1,30 +1,30 @@
 use anyhow::Result;
 use clap::{ArgMatches, Command};
 
-use crate::commands;
+use crate::utils::args::ArgExt;
+
+pub mod list;
 
 macro_rules! each_subcommand {
     ($mac:ident) => {
-        $mac!(react_native_gradle);
-        $mac!(react_native_appcenter);
-        #[cfg(target_os = "macos")]
-        $mac!(react_native_xcode);
+        $mac!(list);
     };
 }
 
 pub fn make_command(mut command: Command) -> Command {
     macro_rules! add_subcommand {
         ($name:ident) => {{
-            command = command.subcommand(commands::$name::make_command(Command::new(
-                &stringify!($name)[13..],
-            )));
+            command = command.subcommand(crate::commands::projects::$name::make_command(
+                Command::new(stringify!($name).replace('_', "-")),
+            ));
         }};
     }
 
     command = command
-        .about("Upload build artifacts for react-native projects.")
+        .about("Manage projects on Sentry.")
         .subcommand_required(true)
-        .arg_required_else_help(true);
+        .arg_required_else_help(true)
+        .org_arg();
     each_subcommand!(add_subcommand);
     command
 }
@@ -32,8 +32,10 @@ pub fn make_command(mut command: Command) -> Command {
 pub fn execute(matches: &ArgMatches) -> Result<()> {
     macro_rules! execute_subcommand {
         ($name:ident) => {{
-            if let Some(sub_matches) = matches.subcommand_matches(&stringify!($name)[13..]) {
-                return commands::$name::execute(&sub_matches);
+            if let Some(sub_matches) =
+                matches.subcommand_matches(&stringify!($name).replace('_', "-"))
+            {
+                return crate::commands::projects::$name::execute(&sub_matches);
             }
         }};
     }
