+++ src/commands/releases/delete.rs
@@ -1,0 +1,28 @@
+use anyhow::Result;
+use clap::{ArgMatches, Command};
+
+use crate::api::Api;
+use crate::config::Config;
+use crate::utils::args::ArgExt;
+
+pub fn make_command(command: Command) -> Command {
+    command.about("Delete a release.").version_arg()
+}
+
+pub fn execute(matches: &ArgMatches) -> Result<()> {
+    let config = Config::current();
+    let api = Api::current();
+    let version = matches.value_of("version").unwrap();
+    let project = config.get_project(matches).ok();
+
+    if api.delete_release(&config.get_org(matches)?, project.as_deref(), version)? {
+        println!("Deleted release {}!", version);
+    } else {
+        println!(
+            "Did nothing. Release with this version ({}) does not exist.",
+            version
+        );
+    }
+
+    Ok(())
+}
